Move main class and adapt readme
